# More efficient encoding of product types