Below are throughput results of benchmarks for encoding of product types using Scala 3 and JDK 21 on on Intel® Core™ i7-11800H CPU @ 2.3GHz (max 4.6GHz).

#### Before
```txt
[info] Benchmark                                      Mode  Cnt       Score       Error  Units
[info] GeoJSONBenchmarks.encodeZio                   thrpt    5   16105.304 ±   289.126  ops/s
[info] GoogleMapsAPIBenchmarks.encodeZio             thrpt    5   21405.929 ±   277.589  ops/s
[info] SyntheticBenchmarks.encodeZio                 thrpt    5  416011.468 ±  1860.251  ops/s
[info] TwitterAPIBenchmarks.encodeZio                thrpt    5   21569.984 ±   162.640  ops/s
```

#### After
```txt
[info] Benchmark                                      Mode  Cnt       Score      Error  Units
[info] GeoJSONBenchmarks.encodeZio                   thrpt    5   16492.856 ±  168.934  ops/s
[info] GoogleMapsAPIBenchmarks.encodeZio             thrpt    5   23977.509 ± 1471.230  ops/s
[info] SyntheticBenchmarks.encodeZio                 thrpt    5  473714.207 ± 6140.838  ops/s
[info] TwitterAPIBenchmarks.encodeZio                thrpt    5   23682.642 ±  238.536  ops/s
```